Updates to token interaction groups and user interface
Allow both fixed and random token events to be assigned to fixed or random tile blocks at any time.
Add checkbox to events that can be part of token interaction groups. Checkbox sets the isReusable property to true, which allows the token event to be placed in multiple tile blocks instead of just used once.